+++ drlwe/relin_key_gen.go
@@ -20,10 +20,10 @@ type RelinearizationKeyGenerator interface {
 
 // RKGProtocol is the structure storing the parameters and and precomputations for the collective relinearization key generation protocol.
 type RKGProtocol struct {
-	ringQModCount   uint64
-	ringQPModCount  uint64
-	alpha           uint64
-	beta            uint64
+	ringQModCount   int
+	ringQPModCount  int
+	alpha           int
+	beta            int
 	ringP           *ring.Ring
 	ringQP          *ring.Ring
 	gaussianSampler *ring.GaussianSampler
@@ -39,13 +39,13 @@ type RKGShare struct {
 }
 
 // NewRKGProtocol creates a new RKG protocol struct
-func NewRKGProtocol(n uint64, q, p []uint64, ephSkPr, sigma float64) *RKGProtocol {
+func NewRKGProtocol(n int, q, p []uint64, ephSkPr, sigma float64) *RKGProtocol {
 	rkg := new(RKGProtocol)
-	rkg.ringQModCount = uint64(len(q))
-	rkg.alpha = uint64(len(p))
+	rkg.ringQModCount = len(q)
+	rkg.alpha = len(p)
 	rkg.ringQPModCount = rkg.ringQModCount + rkg.alpha
 	if rkg.alpha != 0 {
-		rkg.beta = uint64(math.Ceil(float64(len(q)) / float64(len(p))))
+		rkg.beta = int(math.Ceil(float64(len(q)) / float64(len(p))))
 	} else {
 		rkg.beta = 1
 	}
@@ -62,7 +62,7 @@ func NewRKGProtocol(n uint64, q, p []uint64, ephSkPr, sigma float64) *RKGProtoco
 	if err != nil {
 		panic(err) // TODO error
 	}
-	rkg.gaussianSampler = ring.NewGaussianSampler(prng, rkg.ringQP, sigma, uint64(6*sigma))
+	rkg.gaussianSampler = ring.NewGaussianSampler(prng, rkg.ringQP, sigma, int(6*sigma))
 	rkg.ternarySampler = ring.NewTernarySampler(prng, rkg.ringQP, ephSkPr, true)
 	rkg.tmpPoly1, rkg.tmpPoly2 = rkg.ringQP.NewPoly(), rkg.ringQP.NewPoly()
 	return rkg
@@ -74,7 +74,7 @@ func (ekg *RKGProtocol) AllocateShares() (ephSk *rlwe.SecretKey, r1 *RKGShare, r
 	r1, r2 = new(RKGShare), new(RKGShare)
 	r1.value = make([][2]*ring.Poly, ekg.beta)
 	r2.value = make([][2]*ring.Poly, ekg.beta)
-	for i := uint64(0); i < ekg.beta; i++ {
+	for i := 0; i < ekg.beta; i++ {
 		r1.value[i][0] = ekg.ringQP.NewPoly()
 		r1.value[i][1] = ekg.ringQP.NewPoly()
 		r2.value[i][0] = ekg.ringQP.NewPoly()
@@ -96,19 +96,19 @@ func (ekg *RKGProtocol) GenShareRoundOne(sk *rlwe.SecretKey, crp []*ring.Poly, e
 	ekg.ternarySampler.Read(ephSkOut.Value)
 	ekg.ringQP.NTT(ephSkOut.Value, ephSkOut.Value)
 
-	for i := uint64(0); i < ekg.beta; i++ {
+	for i := 0; i < ekg.beta; i++ {
 		// h = e
 		ekg.gaussianSampler.Read(shareOut.value[i][0])
 		ekg.ringQP.NTT(shareOut.value[i][0], shareOut.value[i][0])
 
 		// h = sk*CrtBaseDecompQi + e
-		for j := uint64(0); j < ekg.alpha; j++ {
+		for j := 0; j < ekg.alpha; j++ {
 			index := i*ekg.alpha + j
 			qi := ekg.ringQP.Modulus[index]
 			skP := ekg.tmpPoly1.Coeffs[index]
 			h := shareOut.value[i][0].Coeffs[index]
 
-			for w := uint64(0); w < ekg.ringQP.N; w++ {
+			for w := 0; w < ekg.ringQP.N; w++ {
 				h[w] = ring.CRed(h[w]+skP[w], qi)
 			}
 
@@ -145,7 +145,7 @@ func (ekg *RKGProtocol) GenShareRoundTwo(ephSk, sk *rlwe.SecretKey, round1 *RKGS
 
 	// Each sample is of the form [-u*a_i + s*w_i + e_i]
 	// So for each element of the base decomposition w_i :
-	for i := uint64(0); i < ekg.beta; i++ {
+	for i := 0; i < ekg.beta; i++ {
 
 		// Computes [(sum samples)*sk + e_1i, sk*a + e_2i]
 
@@ -169,7 +169,7 @@ func (ekg *RKGProtocol) GenShareRoundTwo(ephSk, sk *rlwe.SecretKey, round1 *RKGS
 // AggregateShares combines two RKG shares into a single one
 func (ekg *RKGProtocol) AggregateShares(share1, share2, shareOut *RKGShare) {
 
-	for i := uint64(0); i < ekg.beta; i++ {
+	for i := 0; i < ekg.beta; i++ {
 		ekg.ringQP.Add(share1.value[i][0], share2.value[i][0], shareOut.value[i][0])
 		ekg.ringQP.Add(share1.value[i][1], share2.value[i][1], shareOut.value[i][1])
 	}
@@ -177,7 +177,7 @@ func (ekg *RKGProtocol) AggregateShares(share1, share2, shareOut *RKGShare) {
 
 // GenRelinearizationKey computes the generated RLK from the public shares and write the result in evalKeyOut
 func (ekg *RKGProtocol) GenRelinearizationKey(round1 *RKGShare, round2 *RKGShare, evalKeyOut *rlwe.RelinearizationKey) {
-	for i := uint64(0); i < ekg.beta; i++ {
+	for i := 0; i < ekg.beta; i++ {
 		ekg.ringQP.Add(round2.value[i][0], round2.value[i][1], evalKeyOut.Keys[0].Value[i][0])
 		evalKeyOut.Keys[0].Value[i][1].Copy(round1.value[i][1])
 
@@ -190,7 +190,7 @@ func (ekg *RKGProtocol) GenRelinearizationKey(round1 *RKGShare, round2 *RKGShare
 func (share *RKGShare) MarshalBinary() ([]byte, error) {
 	//we have modulus * bitLog * Len of 1 ring rings
 	rLength := (share.value[0])[0].GetDataLen(true)
-	data := make([]byte, 1+2*rLength*uint64(len(share.value)))
+	data := make([]byte, 1+2*rLength*len(share.value))
 	if len(share.value) > 0xFF {
 		return []byte{}, errors.New("RKGShare : uint8 overflow on length")
 	}
@@ -198,7 +198,7 @@ func (share *RKGShare) MarshalBinary() ([]byte, error) {
 
 	//write all of our rings in the data.
 	//write all the polys
-	ptr := uint64(1)
+	ptr := 1
 	for _, elem := range share.value {
 		_, err := elem[0].WriteTo(data[ptr : ptr+rLength])
 		if err != nil {
