+++ examples/dbfv/pir/main.go
@@ -292,7 +292,7 @@ func rkgphase(params *bfv.Parameters, crsGen *ring.UniformSampler, P []*party) *
 	}
 
 	crp := make([]*ring.Poly, params.Beta()) // for the relinearization keys
-	for i := uint64(0); i < params.Beta(); i++ {
+	for i := 0; i < params.Beta(); i++ {
 		crp[i] = crsGen.ReadNew()
 	}
 
@@ -343,7 +343,7 @@ func rtkphase(params *bfv.Parameters, crsGen *ring.UniformSampler, P []*party) *
 
 	crpRot := make([]*ring.Poly, params.Beta()) // for the rotation keys
 
-	for i := uint64(0); i < params.Beta(); i++ {
+	for i := 0; i < params.Beta(); i++ {
 		crpRot[i] = crsGen.ReadNew()
 	}
